+++ tests/Services/SchematicTest.php
@@ -4,6 +4,7 @@ namespace NerdsAndCompany\Schematic\Services;
 
 use Craft\BaseTest;
 use Craft\AssetSourcesService;
+use Craft\AssetTransformsService;
 use Craft\CategoriesService;
 use Craft\Craft;
 use Craft\FieldsService;
@@ -206,6 +207,20 @@ class SchematicTest extends BaseTest
     }
 
     /**
+     * @return Mock|AssetTransformsService
+     */
+    public function getMockAssetTransformsService()
+    {
+        $mock = $this->getMockBuilder(AssetTransformsService::class)
+            ->disableOriginalConstructor()
+            ->getMock();
+
+        $mock->expects($this->exactly(1))->method('getAllTransforms')->willReturn([]);
+
+        return $mock;
+    }
+
+    /**
      * @return Mock|CategoriesService
      */
     public function getMockCategoriesService()
@@ -240,6 +255,7 @@ class SchematicTest extends BaseTest
     {
         $this->createMockService(Locales::class, 'schematic_locales');
         $this->createMockService(AssetSources::class, 'schematic_assetSources');
+        $this->createMockService(AssetTransforms::class, 'schematic_assetTransforms');
         $this->createMockService(Fields::class, 'schematic_fields');
         $this->createMockService(GlobalSets::class, 'schematic_globalSets');
         $this->createMockService(Plugins::class, 'schematic_plugins');
@@ -297,6 +313,9 @@ class SchematicTest extends BaseTest
 
         $mockAssetSourcesService = $this->getMockAssetSourcesService();
         $this->setCraftComponent('assetSources', $mockAssetSourcesService);
+
+        $mockAssetTransformsService = $this->getMockAssetTransformsService();
+        $this->setCraftComponent('assetTransforms', $mockAssetTransformsService);
 
         $mockCategoriesService = $this->getMockCategoriesService();
         $this->setCraftComponent('categories', $mockCategoriesService);